todo completo, falta busqueda

+++ main.js
@@ -72,43 +72,51 @@ const form = new Vue({
             fecha: '',
             email: '',
             armas: '',
-            data: [1,2],
-            idEliminar: null,
+            data: [],
         };
     },
     mounted(){
-        this.data = [];
-        if(localStorage.length > 0){
-            this.data =  JSON.parse(localStorage.getItem('data'));       
-        }
+        
     },
     methods: {
         saveData(e){
             e.preventDefault();
-            if(!this.idExists(this.codigo)){
-                let obj = {
-                    codigo: this.codigo,
-                    nombre: this.nombre,
-                    genero: this.genero,
-                    fecha: this.fecha,
-                    email: this.email,
-                    armas: this.armas
+            this.data = [];
+            if(localStorage.length > 0){
+                this.data =  JSON.parse(localStorage.getItem('data'));       
+            }
+
+            if(!this.idExists(this.codigo) && this.notEmptyValues()){
+                if(this.tryParseInt(this.codigo, false) != false && this.hasOnlyLetters(this.nombre) && this.hasOnlyLetters(this.genero) && this.hasOnlyLetters(this.armas) && this.isEmailValid(this.email) && this.isDateValid(this.fecha)){
+                    let obj = {
+                        codigo: this.codigo,
+                        nombre: this.nombre,
+                        genero: this.genero,
+                        fecha: this.fecha,
+                        email: this.email,
+                        armas: this.armas
+                    }
+    
+                    console.log(obj);
+                    this.data.push(obj);
+    
+                    localStorage.setItem('data', JSON.stringify(this.data));
+                    this.codigo = null;
+                    this.nombre = null;
+                    this.genero = null;
+                    this.fecha = null;
+                    this.email = null;
+                    this.armas = null;
+                    
+                    //Asignar valor a la propiedad data del componente table para que se actualice
+                    table.data = this.data;
                 }
-
-                console.log(obj);
-                console.log(this.data);
-                this.data.push(obj);
-
-                localStorage.setItem('data', JSON.stringify(this.data));
-                this.codigo = null;
-                this.nombre = null;
-                this.genero = null;
-                this.fecha = null;
-                this.email = null;
-                this.armas = null;
+                else{
+                    console.log('Los campos no se han llenado correctamente')
+                }
             }
             else
-                console.log('El id ya existe');
+                console.log('Valor inválido o faltante');
         },
         idExists(codigo){
             let datos =  JSON.parse(localStorage.getItem('data'));       
@@ -119,27 +127,120 @@ const form = new Vue({
             }
 
             console.log(ids);
-            if(ids.includes(codigo)) 
+            if(ids.includes(codigo)){
+                console.log('El código ya existe');
                 return true
+            }
             else 
                 return false
         },
+        notEmptyValues(){
+            try{
+                if(this.codigo.trim().length > 0 && this.nombre.trim().length > 0 && this.genero.trim().length > 0 && this.fecha.trim().length > 0 && this.email.trim().length > 0 && this.armas.trim().length > 0 && this.isEmailValid(this.email))
+                    return true
+                else 
+                    return false
+            }
+            catch(err){
+                console.log('Hacen falta campos por completar');
+            }
+        },
+        tryParseInt(str,defaultValue) {
+            var retValue = defaultValue;
+            if(str !== null) {
+                if(str.length > 0) {
+                    if (!isNaN(str)) {
+                        retValue = parseInt(str);
+                    }
+                }
+            }
+            return retValue;
+       },
+       hasOnlyLetters(value){
+           //Regular expression for names
+            if(/^[a-zA-Z]+(([',. -][a-zA-Z ])?[a-zA-Z]*)*$/.test(value) == true)
+                return true;
+            else
+                return false;
+       },
+       isEmailValid(value){
+           //RegExp from https://emailregex.com/
+           
+           if(/^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/.test(value))
+            return true;
+           else
+            return false;
+       },
+       isDateValid(value){
+           //RegExp from https://www.regular-expressions.info/dates.html
+
+           if(/^(19|20)\d\d[- /.](0[1-9]|1[012])[- /.](0[1-9]|[12][0-9]|3[01])$/.test(value))
+            return true;
+           else
+            return false;
+       }
+    }
+});
+
+//Tabla
+const table = new Vue({
+    el: '#tabla',
+    data(){
+        return{
+            data: JSON.parse(localStorage.getItem('data'))
+        }
+    },
+});
+
+const formEliminar = new Vue({
+    el: '#formEliminar',
+    data(){
+        return{
+            idEliminar: null
+        }
+    },
+    methods: {
         deleteRow(e){
             e.preventDefault();
             if(this.idExists(this.idEliminar)){
-                this.data = JSON.parse(localStorage.getItem('data')); 
+                let data = JSON.parse(localStorage.getItem('data')); 
                 //Eliminar el elemento indicado
                 //indexOf devuelve posición
                 //splice elimina el elemento
-                this.data.splice(this.data.indexOf(this.idEliminar), 1);
-                localStorage.setItem('data', JSON.stringify(this.data));
+                data.splice(this.getIndexNum(this.idEliminar), 1);
+                localStorage.setItem('data', JSON.stringify(data));
+                table.data = data;
                 this.idEliminar = null;
             }
             else 
                 console.log('El valor no existe');
+        },
+        idExists(codigo){
+            let datos =  JSON.parse(localStorage.getItem('data'));       
+            let ids = [];
+
+            for(let i in datos){
+                ids.push(datos[i].codigo);    
+            }
+
+            console.log(ids);
+            if(ids.includes(codigo)){
+                console.log('El código ya existe');
+                return true
+            }
+            else 
+                return false
+        },
+        getIndexNum(value){
+            let datos =  JSON.parse(localStorage.getItem('data'));       
+            let ids = [];
+
+            for(let i in datos){
+                ids.push(datos[i].codigo);    
+            }
+
+            return ids.indexOf(value)
         }
     }
-});
-
-
+})
